+++ src/xml.h
@@ -1,15 +1,10 @@
 #ifndef XML_H
 #define XML_H
 
-#include <stdbool.h>
+#include "string_view.h"
 #include <stdio.h>
 
 typedef enum { XML_TOKEN_TEXT, XML_TOKEN_NODE } XML_TokenType;
-
-typedef struct {
-    const char *start;
-    size_t length;
-} XML_StringView;
 
 typedef struct {
     XML_StringView name;
@@ -50,8 +45,5 @@ bool XML_get_attribute(XML_Token token, const char *property,
 
 char *XML_read_file(const char *file_name);
 void XML_debug_print(FILE *file, XML_Token root);
-bool XML_str_eq(XML_StringView a, XML_StringView b);
-bool XML_str_eq_cstr(XML_StringView a, const char *b);
-size_t XML_strlen(const char *str);
 
 #endif
